More Harmony patch reorg
