+++ Source/RP0.csproj
@@ -61,23 +61,29 @@
     <Compile Include="ConfidenceWidget.cs" />
     <Compile Include="Confidence.cs" />
     <Compile Include="Debug.cs" />
-    <Compile Include="Harmony\AdminPatcher.cs" />
+    <Compile Include="Harmony\Administration.cs" />
+    <Compile Include="Harmony\RDController.cs" />
+    <Compile Include="Harmony\PartListTooltip.cs" />
+    <Compile Include="Harmony\RDPartList.cs" />
+    <Compile Include="Harmony\ScienceWidget.cs" />
+    <Compile Include="Harmony\FundsWidget.cs" />
+    <Compile Include="Harmony\RealFuels.cs" />
     <Compile Include="UnlockSubsidyHandler.cs" />
     <Compile Include="LocalizationHandler.cs" />
     <Compile Include="EventRegisterer.cs" />
     <Compile Include="EntryCostStorage.cs" />
     <Compile Include="FirstStart.cs" />
-    <Compile Include="Harmony\AppLauncherPatcher.cs" />
+    <Compile Include="Harmony\ApplicationLauncher.cs" />
     <Compile Include="Harmony\KerbalismPatcher.cs" />
-    <Compile Include="Harmony\ContractsAppPatcher.cs" />
-    <Compile Include="Harmony\DifficultyPatcher.cs" />
+    <Compile Include="Harmony\ContractsApp.cs" />
+    <Compile Include="Harmony\DifficultyOptionsMenu.cs" />
     <Compile Include="Harmony\CurrencyWidgetPatcher.cs" />
-    <Compile Include="Harmony\ReputationPatcher.cs" />
+    <Compile Include="Harmony\Reputation.cs" />
     <Compile Include="Harmony\ContractPatcher.cs" />
     <Compile Include="Harmony\RnDPatcher.cs" />
     <Compile Include="GameplayTips.cs" />
     <Compile Include="Harmony\HarmonyPatcher.cs" />
-    <Compile Include="Harmony\CBKPatcher.cs" />
+    <Compile Include="Harmony\CustomBarnKit.cs" />
     <Compile Include="LoadingScreenChanger.cs" />
     <Compile Include="Maintenance\MaintenanceSettings.cs" />
     <Compile Include="Maintenance\MaintenanceHandler.cs" />
@@ -100,7 +106,7 @@
     <Compile Include="Avionics\ProceduralAvionicsTechNode.cs" />
     <Compile Include="Avionics\ProceduralAvionicsUnlockedNodes.cs" />
     <Compile Include="Avionics\ProceduralAvionicsUtils.cs" />
-    <Compile Include="Harmony\StrategyPatches.cs" />
+    <Compile Include="Harmony\StrategySystem.cs" />
     <Compile Include="Programs\ProgramSpeedListItem.cs" />
     <Compile Include="Programs\AdminExtender.cs" />
     <Compile Include="Programs\ProgramStrategy.cs" />
